+++ python/pdstools/infinity/resources/prediction_studio/v25/prediction_studio/__init__.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
 from ._async import AsyncPredictionStudio
-from ._mixin import _PredictionStudiov26Mixin
+from ._mixin import _PredictionStudiov25Mixin
 from ._sync import PredictionStudio
 
 __all__ = [
     "AsyncPredictionStudio",
     "PredictionStudio",
-    "_PredictionStudiov26Mixin",
+    "_PredictionStudiov25Mixin",
 ]